requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
redis
celery
python-dotenv
stripe
asyncpg